+++ Runtime/Singleton/GenericDataManager.cs
@@ -7,7 +7,8 @@ using UnityEngine;
 namespace GameUtils
 {
     [DeclareBoxGroup("debug", Title = "Debug")]
-    [DefaultExecutionOrder(-100)]
+    // Ensures data manager singletons initialize much earlier than standard MonoBehaviours.
+    [DefaultExecutionOrder(-10000)]
     public abstract class GenericDataManager<T1, T2> : MonoBehaviour, ILoggable where T1 : GenericDataManager<T1, T2> where T2 : ItemIdentifierData
     {
         [SerializeField, Group("debug")] private bool _logEnabled = true;