+++ Runtime/UI/Loading/LoadingManager.cs
@@ -5,7 +5,8 @@ using UnityEngine;
 
 namespace GameUtils
 {
-    [DefaultExecutionOrder(0)]
+    // Ensures this singleton initializes much earlier than standard MonoBehaviours.
+    [DefaultExecutionOrder(-10000)]
     public class LoadingManager : Singleton<LoadingManager>
     {
         [Tab("References")]